+++ db_conn.py
@@ -1,0 +1,4 @@
+import pymongo
+uri = "mongodb+srv://root:root920412@cluster0.lvs2gvu.mongodb.net/?retryWrites=true&w=majority"
+client = pymongo.MongoClient(uri)
+db = client.member_system
