+++ templates/signup.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <link href='https://unpkg.com/boxicons@2.1.4/css/boxicons.min.css' rel='stylesheet'>
-    <link rel="stylesheet" type ="text/css" href="signup.css" />
+    <link rel="stylesheet" type="text/css" href="{{ url_for('static', filename='signup.css') }}">
     <style>
       .radio-container {
         display: flex;
@@ -21,7 +21,7 @@
       <div id="container2">
 
         <h3 align="center">註冊 Sign Up</h3>
-        <form action="/signup" method="POST">
+        <form action="{{ url_for('sign.signup') }}" method="POST">
           <div class="input-box">
             <input type="text" id="fullname" name="fullname" placeholder="使用者全名" required>
             <div class="tab"></div>
